Moved testing framework from unittest to pytest
