+++ .github/workflows/python-app.yml
@@ -34,14 +34,6 @@ jobs:
         python3 -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
-    - name: set pythonpath
+    - name: Run all tests using pytest for Macro-Zero Application
       run: |
-        echo "PYTHONPATH=/home/runner/work/macro-zero/macro-zero/src/" >> $GITHUB_ENV
-
-    - name: Run all unit tests for mkeyboard module
-      run: |
-        python3 -m unittest discover -v -s src/tests/unit/mkeyboard/
-
-    - name: Run all unit tests for app
-      run: |
-        python3 -m unittest discover -v -s src/tests/unit/app/
+        python3 -m pytest --cov-report html:cov_html --cov=src tests/
