+++ requirements.txt
@@ -3,3 +3,5 @@ Pillow>=8.2.0
 spidev~=3.5
 Flask~=2.0.1
 smbus2>=0.4.1
+pytest>=6.2.4
+pytest-cov>=2.12.1
