+++ src/modules/__init__.py
@@ -38,7 +38,7 @@ RE_SW_PIN = 11
 RE_DR_PIN = 15
 RE_CLK_PIN = 13
 
-__version__ = "0.010.1"
+__version__ = "0.015.2"
 
 path = os.path.dirname(__file__)
 """Path to modules package directory."""
