+++ tests/unit/mkeyboard/test_verify_report.py
@@ -1,10 +1,5 @@
 import unittest
-import sys
-import os
-
-sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.getcwd()))))
-
-from macrozero import MacroZero
+from src.macrozero import MacroZero
 
 
 class VerifyReportTestCase(unittest.TestCase):
